knapsack + GLPK

+++ src/com/company/knapsack/Item.java
@@ -1,19 +1,19 @@
 package com.company.knapsack;
 
 public class Item {
-    private final double weight;
-    private final double value;
+    private final int weight;
+    private final int value;
 
-    public Item(double weight, double value) {
+    public Item(int weight, int value) {
         this.weight = weight;
         this.value = value;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
